django-DRF/requirements.txt
asgiref==3.5.0
attrs==22.1.0
Django==4.0.2
djangorestframework==3.13.1
iniconfig==1.1.1
mysqlclient==2.1.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.9
pytest==7.1.2
pytest-django==4.5.2
pytz==2021.3
sqlparse==0.4.2
tomli==2.0.1
tzdata==2021.5
